+++ tests/tests_unit/rules/test_importers/test_owl_importer.py
@@ -4,7 +4,9 @@ from cognite.neat.rules.models._rules._types import ClassEntity, EntityTypes
 
 
 def test_owl_importer():
-    rules, _ = importers.OWLImporter(owl_filepath="https://data.nobelprize.org/terms.rdf").to_rules()
+    rules, _ = importers.OWLImporter(
+        owl_filepath="https://data.nobelprize.org/terms.rdf", make_compliant=True
+    ).to_rules()
 
     assert len(rules.classes) == 11
     assert len(rules.properties) == 16